improve: changes short variables to int

+++ Scripts/Gameplay/DamageableEntity_DamageHitNotify.cs
@@ -24,7 +24,7 @@ namespace MultiplayerARPG
             int damage,
             CharacterItem weapon,
             BaseSkill skill,
-            short skillLevel,
+            int skillLevel,
             CharacterBuff buff,
             bool isDamageOverTime)
         {
